[Decomp][Player] Cohesion split: extract save system to PlayerSave.cpp
Moves the entire SAVE SYSTEM block plus three scattered save helpers
out of Player.cpp into a new file under src/game/Object/.

Player.cpp: 19,947 -> 18,889 lines (-1,058).
PlayerSave.cpp: 1,141 lines, 18 methods.

Methods moved:
- SaveToDB, SaveInventoryAndGoldToDB, SaveGoldToDB
- _SaveActions, _SaveAuras, _SaveInventory, _SaveMail
- _SaveQuestStatus, _SaveDailyQuestStatus, _SaveWeeklyQuestStatus,
  _SaveMonthlyQuestStatus
- _SaveSkills, _SaveSpells, _SaveTalents, _SaveStats
- SavePositionInDB, _SaveEquipmentSets, _SaveBGData

No header change; no signature change; no behaviour change. Symmetric
to PR #145 (PlayerLoad.cpp) and picked up automatically by the
file(GLOB Object/*.cpp) rule in src/game/CMakeLists.txt.
