+++ src/game/Object/Player.cpp
@@ -11487,930 +11487,6 @@ bool Player::isAllowedToLoot(Creature* creature)
     }
 }
 
-/*********************************************************/
-/***                   SAVE SYSTEM                     ***/
-/*********************************************************/
-
-void Player::SaveToDB()
-{
-    // we should assure this: ASSERT((m_nextSave != sWorld.getConfig(CONFIG_UINT32_INTERVAL_SAVE)));
-    // delay auto save at any saves (manual, in code, or autosave)
-    m_nextSave = sWorld.getConfig(CONFIG_UINT32_INTERVAL_SAVE);
-
-    // lets allow only players in world to be saved
-    if (IsBeingTeleportedFar())
-    {
-        ScheduleDelayedOperation(DELAYED_SAVE_PLAYER);
-        return;
-    }
-
-    // first save/honor gain after midnight will also update the player's honor fields
-    UpdateHonorKills();
-
-    DEBUG_FILTER_LOG(LOG_FILTER_PLAYER_STATS, "The value of player %s at save: ", m_name.c_str());
-    outDebugStatsValues();
-
-    CharacterDatabase.BeginTransaction();
-
-#ifdef ENABLE_ELUNA
-    // Hack to check that this is not on create save
-    if (Eluna* e = GetEluna())
-    {
-        if (!HasAtLoginFlag(AT_LOGIN_FIRST))
-        {
-            e->OnSave(this);
-        }
-    }
-#endif /* ENABLE_ELUNA */
-
-    static SqlStatementID delChar ;
-    static SqlStatementID insChar ;
-
-    SqlStatement stmt = CharacterDatabase.CreateStatement(delChar, "DELETE FROM `characters` WHERE `guid` = ?");
-    stmt.PExecute(GetGUIDLow());
-
-    SqlStatement uberInsert = CharacterDatabase.CreateStatement(insChar, "INSERT INTO `characters` (`guid`,`account`,`name`,`race`,`class`,`gender`, "
-                              "`level`,`xp`,`money`,`playerBytes`,`playerBytes2`,`playerFlags`,"
-                              "`map`, `dungeon_difficulty`, `position_x`, `position_y`, `position_z`, `orientation`, "
-                              "`taximask`, `online`, `cinematic`, "
-                              "`totaltime`, `leveltime`, `rest_bonus`, `logout_time`, `is_logout_resting`, `resettalents_cost`, `resettalents_time`, `primary_trees`, "
-                              "`trans_x`, `trans_y`, `trans_z`, `trans_o`, `transguid`, `extra_flags`, `stable_slots`, `at_login`, `zone`, "
-                              "`death_expire_time`, `taxi_path`, `totalKills`, "
-                              "`todayKills`, `yesterdayKills`, `chosenTitle`, "
-                              "`watchedFaction`, `drunk`, `health`, `power1`, `power2`, `power3`, "
-                              "`power4`, `power5`, `specCount`, `activeSpec`, `exploredZones`, `equipmentCache`, `knownTitles`, "
-                              "`actionBars`, `slot`,   `deleteInfos_Account`, `deleteInfos_Name`, `deleteDate`, `createdDate`) "
-                              "VALUES (?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, "
-                              "?, ?, ?, ?, ?, ?, "
-                              "?, ?, ?, "
-                              "?, ?, ?, ?, ?, ?, ?, ?, "
-                              "?, ?, ?, ?, ?, ?, ?, ?, ?, "
-                              "?, ?, ?, "
-                              "?, ?, ?, ?, ?, ?, ?, ?, ?, "
-                              "?, ?, ?, ?, ?, ?, ?, "
-                              "?, ?, null, null, null, ?) ");
-
-    uberInsert.addUInt32(GetGUIDLow());
-    uberInsert.addUInt32(GetSession()->GetAccountId());
-    uberInsert.addString(m_name.c_str());
-    uberInsert.addUInt8(getRace());
-    uberInsert.addUInt8(getClass());
-    uberInsert.addUInt8(getGender());
-    uberInsert.addUInt32(getLevel());
-    uberInsert.addUInt32(GetUInt32Value(PLAYER_XP));
-    uberInsert.addUInt64(GetMoney());
-    uberInsert.addUInt32(GetUInt32Value(PLAYER_BYTES));
-    uberInsert.addUInt32(GetUInt32Value(PLAYER_BYTES_2));
-    uberInsert.addUInt32(GetUInt32Value(PLAYER_FLAGS));
-
-    if (!IsBeingTeleported())
-    {
-        uberInsert.addUInt32(GetMapId());
-        uberInsert.addUInt32(uint32(GetDungeonDifficulty()));
-        uberInsert.addFloat(finiteAlways(GetPositionX()));
-        uberInsert.addFloat(finiteAlways(GetPositionY()));
-        uberInsert.addFloat(finiteAlways(GetPositionZ()));
-        uberInsert.addFloat(finiteAlways(GetOrientation()));
-    }
-    else
-    {
-        uberInsert.addUInt32(GetTeleportDest().mapid);
-        uberInsert.addUInt32(uint32(GetDungeonDifficulty()));
-        uberInsert.addFloat(finiteAlways(GetTeleportDest().coord_x));
-        uberInsert.addFloat(finiteAlways(GetTeleportDest().coord_y));
-        uberInsert.addFloat(finiteAlways(GetTeleportDest().coord_z));
-        uberInsert.addFloat(finiteAlways(GetTeleportDest().orientation));
-    }
-
-    std::ostringstream ss;
-    ss << m_taxi;                                   // string with TaxiMaskSize numbers
-    uberInsert.addString(ss);
-
-    uberInsert.addUInt32(IsInWorld() ? 1 : 0);
-
-    uberInsert.addUInt32(m_cinematic);
-
-    uberInsert.addUInt32(m_Played_time[PLAYED_TIME_TOTAL]);
-    uberInsert.addUInt32(m_Played_time[PLAYED_TIME_LEVEL]);
-
-    uberInsert.addFloat(finiteAlways(m_rest_bonus));
-    uberInsert.addUInt64(uint64(time(NULL)));
-    uberInsert.addUInt32(HasFlag(PLAYER_FLAGS, PLAYER_FLAGS_RESTING) ? 1 : 0);
-    // save, far from tavern/city
-    // save, but in tavern/city
-    uberInsert.addUInt32(m_resetTalentsCost);
-    uberInsert.addUInt64(uint64(m_resetTalentsTime));
-    ss.str("");
-    for (int i = 0; i < MAX_TALENT_SPEC_COUNT; ++i)
-    {
-        ss << m_talentsPrimaryTree[i] << " ";
-    }
-    uberInsert.addString(ss);
-
-    uberInsert.addFloat(finiteAlways(m_movementInfo.GetTransportPos()->x));
-    uberInsert.addFloat(finiteAlways(m_movementInfo.GetTransportPos()->y));
-    uberInsert.addFloat(finiteAlways(m_movementInfo.GetTransportPos()->z));
-    uberInsert.addFloat(finiteAlways(m_movementInfo.GetTransportPos()->o));
-    if (m_transport)
-    {
-        uberInsert.addUInt32(m_transport->GetGUIDLow());
-    }
-    else
-    {
-        uberInsert.addUInt32(0);
-    }
-
-    uberInsert.addUInt32(m_ExtraFlags);
-
-    uberInsert.addUInt32(uint32(m_petMgr.GetStableSlots()));   // to prevent save uint8 as char
-
-    uberInsert.addUInt32(uint32(m_atLoginFlags));
-
-    uberInsert.addUInt32(IsInWorld() ? GetZoneId() : GetCachedZoneId());
-
-    uberInsert.addUInt64(uint64(m_deathExpireTime));
-
-    ss << m_taxi.SaveTaxiDestinationsToString();       // string
-    uberInsert.addString(ss);
-
-    uberInsert.addUInt32(GetUInt32Value(PLAYER_FIELD_LIFETIME_HONORABLE_KILLS));
-
-    uberInsert.addUInt16(GetUInt16Value(PLAYER_FIELD_KILLS, 0));
-
-    uberInsert.addUInt16(GetUInt16Value(PLAYER_FIELD_KILLS, 1));
-
-    uberInsert.addUInt32(GetUInt32Value(PLAYER_CHOSEN_TITLE));
-
-    // FIXME: at this moment send to DB as unsigned, including unit32(-1)
-    uberInsert.addUInt32(GetUInt32Value(PLAYER_FIELD_WATCHED_FACTION_INDEX));
-
-    uberInsert.addUInt8(GetDrunkValue());
-
-    uberInsert.addUInt32(GetHealth());
-
-    for (uint32 i = 0; i < MAX_STORED_POWERS; ++i)  // power1 to power5
-    {
-        uberInsert.addUInt32(GetPowerByIndex(i));
-    }
-
-    uberInsert.addUInt32(uint32(m_specsCount));
-    uberInsert.addUInt32(uint32(m_activeSpec));
-
-    for (uint32 i = 0; i < PLAYER_EXPLORED_ZONES_SIZE; ++i) // string
-    {
-        ss << GetUInt32Value(PLAYER_EXPLORED_ZONES_1 + i) << " ";
-    }
-    uberInsert.addString(ss); // exploredZOnes
-
-    for (uint32 i = 0; i < EQUIPMENT_SLOT_END * 2; ++i)     // string
-    {
-        ss << GetUInt32Value(PLAYER_VISIBLE_ITEM_1_ENTRYID + i) << " ";
-    }
-    uberInsert.addString(ss); // EquipmentCache
-
-    for (uint32 i = 0; i < KNOWN_TITLES_SIZE * 2; ++i)      // string
-    {
-        ss << GetUInt32Value(PLAYER__FIELD_KNOWN_TITLES + i) << " ";
-    }
-    uberInsert.addString(ss);
-
-    uberInsert.addUInt32(uint32(GetByteValue(PLAYER_FIELD_BYTES, 2))); // actionbars
-
-    uberInsert.addUInt8(m_slot);
-
-    uberInsert.addUInt32(GetCreatedDate());
-
-    uberInsert.Execute();
-
-    if (m_mailsUpdated)                                     // save mails only when needed
-    {
-        _SaveMail();
-    }
-
-    _SaveBGData();
-    _SaveInventory();
-    _SaveQuestStatus();
-    _SaveDailyQuestStatus();
-    _SaveWeeklyQuestStatus();
-    _SaveMonthlyQuestStatus();
-    _SaveSpells();
-    _SaveSpellCooldowns();
-    _SaveActions();
-    _SaveAuras();
-    _SaveSkills();
-    m_achievementMgr.SaveToDB();
-    m_reputationMgr.SaveToDB();
-    _SaveCurrencies();
-    _SaveEquipmentSets();
-    GetSession()->SaveTutorialsData();                      // changed only while character in game
-    _SaveGlyphs();
-    _SaveTalents();
-
-    CharacterDatabase.CommitTransaction();
-
-    // check if stats should only be saved on logout
-    // save stats can be out of transaction
-    if (m_session->isLogingOut() || !sWorld.getConfig(CONFIG_BOOL_STATS_SAVE_ONLY_ON_LOGOUT))
-    {
-        _SaveStats();
-    }
-
-    // save pet (hunter pet level and experience and all type pets health/mana).
-    if (Pet* pet = GetPet())
-    {
-        pet->SavePetToDB(PET_SAVE_AS_CURRENT);
-    }
-}
-
-// fast save function for item/money cheating preventing - save only inventory and money state
-void Player::SaveInventoryAndGoldToDB()
-{
-    _SaveInventory();
-    SaveGoldToDB();
-}
-
-/**
- * @brief Persists the player's current money value to the database.
- */
-void Player::SaveGoldToDB()
-{
-    static SqlStatementID updateGold ;
-
-    SqlStatement stmt = CharacterDatabase.CreateStatement(updateGold, "UPDATE `characters` SET `money` = ? WHERE `guid` = ?");
-    stmt.PExecute(GetMoney(), GetGUIDLow());
-}
-
-/**
- * @brief Saves changed action bar bindings to the database.
- */
-void Player::_SaveActions()
-{
-    static SqlStatementID insertAction ;
-    static SqlStatementID updateAction ;
-    static SqlStatementID deleteAction ;
-
-    for (int i = 0; i < MAX_TALENT_SPEC_COUNT; ++i)
-    {
-        for (ActionButtonList::iterator itr = m_actionButtons[i].begin(); itr != m_actionButtons[i].end();)
-        {
-            switch (itr->second.uState)
-            {
-                case ACTIONBUTTON_NEW:
-                {
-                    SqlStatement stmt = CharacterDatabase.CreateStatement(insertAction, "INSERT INTO `character_action` (`guid`,`spec`, `button`,`action`,`type`) VALUES (?, ?, ?, ?, ?)");
-                    stmt.addUInt32(GetGUIDLow());
-                    stmt.addUInt32(i);
-                    stmt.addUInt32(uint32(itr->first));
-                    stmt.addUInt32(itr->second.GetAction());
-                    stmt.addUInt32(uint32(itr->second.GetType()));
-                    stmt.Execute();
-                    itr->second.uState = ACTIONBUTTON_UNCHANGED;
-                    ++itr;
-                }
-                break;
-                case ACTIONBUTTON_CHANGED:
-                {
-                    SqlStatement stmt = CharacterDatabase.CreateStatement(updateAction, "UPDATE `character_action` SET `action` = ?, `type` = ? WHERE `guid` = ? AND `button` = ? AND `spec` = ?");
-                    stmt.addUInt32(itr->second.GetAction());
-                    stmt.addUInt32(uint32(itr->second.GetType()));
-                    stmt.addUInt32(GetGUIDLow());
-                    stmt.addUInt32(uint32(itr->first));
-                    stmt.addUInt32(i);
-                    stmt.Execute();
-                    itr->second.uState = ACTIONBUTTON_UNCHANGED;
-                    ++itr;
-                }
-                break;
-                case ACTIONBUTTON_DELETED:
-                {
-                    SqlStatement stmt = CharacterDatabase.CreateStatement(deleteAction, "DELETE FROM `character_action` WHERE `guid` = ? AND `button` = ? AND `spec` = ?");
-                    stmt.addUInt32(GetGUIDLow());
-                    stmt.addUInt32(uint32(itr->first));
-                    stmt.addUInt32(i);
-                    stmt.Execute();
-                    m_actionButtons[i].erase(itr++);
-                }
-                break;
-                default:
-                    ++itr;
-                    break;
-            }
-        }
-    }
-}
-
-/**
- * @brief Saves eligible active aura state to the database.
- */
-void Player::_SaveAuras()
-{
-    static SqlStatementID deleteAuras ;
-    static SqlStatementID insertAuras ;
-
-    SqlStatement stmt = CharacterDatabase.CreateStatement(deleteAuras, "DELETE FROM `character_aura` WHERE `guid` = ?");
-    stmt.PExecute(GetGUIDLow());
-
-    SpellAuraHolderMap const& auraHolders = GetSpellAuraHolderMap();
-
-    if (auraHolders.empty())
-    {
-        return;
-    }
-
-    stmt = CharacterDatabase.CreateStatement(insertAuras, "INSERT INTO `character_aura` (`guid`, `caster_guid`, `item_guid`, `spell`, `stackcount`, `remaincharges`, "
-            "`basepoints0`, `basepoints1`, `basepoints2`, `periodictime0`, `periodictime1`, `periodictime2`, `maxduration`, `remaintime`, `effIndexMask`) "
-            "VALUES (?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?)");
-
-    for (SpellAuraHolderMap::const_iterator itr = auraHolders.begin(); itr != auraHolders.end(); ++itr)
-    {
-        SpellAuraHolder* holder = itr->second;
-        // skip all holders from spells that are passive or channeled
-        // save singleTarget auras if self cast.
-        bool selfCastHolder = holder->GetCasterGuid() == GetObjectGuid();
-        TrackedAuraType trackedType = holder->GetTrackedAuraType();
-        if (!holder->IsPassive() && !IsChanneledSpell(holder->GetSpellProto()) &&
-           (trackedType == TRACK_AURA_TYPE_NOT_TRACKED || (trackedType == TRACK_AURA_TYPE_SINGLE_TARGET && selfCastHolder)))
-        {
-            int32  damage[MAX_EFFECT_INDEX];
-            uint32 periodicTime[MAX_EFFECT_INDEX];
-            uint32 effIndexMask = 0;
-
-            for (uint32 i = 0; i < MAX_EFFECT_INDEX; ++i)
-            {
-                damage[i] = 0;
-                periodicTime[i] = 0;
-
-                if (Aura* aur = holder->GetAuraByEffectIndex(SpellEffectIndex(i)))
-                {
-                    // don't save not own area auras
-                    if (aur->IsAreaAura() && holder->GetCasterGuid() != GetObjectGuid())
-                    {
-                        continue;
-                    }
-
-                    damage[i] = aur->GetModifier()->m_amount;
-                    periodicTime[i] = aur->GetModifier()->periodictime;
-                    effIndexMask |= (1 << i);
-                }
-            }
-
-            if (!effIndexMask)
-            {
-                continue;
-            }
-
-            stmt.addUInt32(GetGUIDLow());
-            stmt.addUInt64(holder->GetCasterGuid().GetRawValue());
-            stmt.addUInt32(holder->GetCastItemGuid().GetCounter());
-            stmt.addUInt32(holder->GetId());
-            stmt.addUInt32(holder->GetStackAmount());
-            stmt.addUInt8(holder->GetAuraCharges());
-
-            for (uint32 i = 0; i < MAX_EFFECT_INDEX; ++i)
-            {
-                stmt.addInt32(damage[i]);
-            }
-
-            for (uint32 i = 0; i < MAX_EFFECT_INDEX; ++i)
-            {
-                stmt.addUInt32(periodicTime[i]);
-            }
-
-            stmt.addInt32(holder->GetAuraMaxDuration());
-            stmt.addInt32(holder->GetAuraDuration());
-            stmt.addUInt32(effIndexMask);
-            stmt.Execute();
-        }
-    }
-}
-
-// Player::_SaveGlyphs moved to GlyphMgr::Save (2026-05-12); thin delegating wrapper lives inline in Player.h.
-
-/**
- * @brief Saves inventory state changes and queued item records to the database.
- */
-void Player::_SaveInventory()
-{
-    // force items in buyback slots to new state
-    // and remove those that aren't already
-    for (uint8 i = BUYBACK_SLOT_START; i < BUYBACK_SLOT_END; ++i)
-    {
-        Item* item = m_items[i];
-        if (!item || item->GetState() == ITEM_NEW)
-        {
-            continue;
-        }
-
-        static SqlStatementID delInv ;
-        static SqlStatementID delItemInst ;
-
-        SqlStatement stmt = CharacterDatabase.CreateStatement(delInv, "DELETE FROM `character_inventory` WHERE `item` = ?");
-        stmt.PExecute(item->GetGUIDLow());
-
-        stmt = CharacterDatabase.CreateStatement(delItemInst, "DELETE FROM `item_instance` WHERE `guid` = ?");
-        stmt.PExecute(item->GetGUIDLow());
-
-        m_items[i]->FSetState(ITEM_NEW);
-    }
-
-    // update enchantment durations
-    for (EnchantDurationList::const_iterator itr = m_enchantDuration.begin(); itr != m_enchantDuration.end(); ++itr)
-    {
-        itr->item->SetEnchantmentDuration(itr->slot, itr->leftduration);
-    }
-
-    // if no changes
-    if (m_itemUpdateQueue.empty())
-    {
-        return;
-    }
-
-    // do not save if the update queue is corrupt
-    bool error = false;
-    for (size_t i = 0; i < m_itemUpdateQueue.size(); ++i)
-    {
-        Item* item = m_itemUpdateQueue[i];
-        if (!item || item->GetState() == ITEM_REMOVED)
-        {
-            continue;
-        }
-        Item* test = GetItemByPos(item->GetBagSlot(), item->GetSlot());
-
-        if (test == NULL)
-        {
-            sLog.outError("Player(GUID: %u Name: %s)::_SaveInventory - the bag(%d) and slot(%d) values for the item with guid %d are incorrect, the player doesn't have an item at that position!", GetGUIDLow(), GetName(), item->GetBagSlot(), item->GetSlot(), item->GetGUIDLow());
-            error = true;
-        }
-        else if (test != item)
-        {
-            sLog.outError("Player(GUID: %u Name: %s)::_SaveInventory - the bag(%d) and slot(%d) values for the item with guid %d are incorrect, the item with guid %d is there instead!", GetGUIDLow(), GetName(), item->GetBagSlot(), item->GetSlot(), item->GetGUIDLow(), test->GetGUIDLow());
-            error = true;
-        }
-    }
-
-    if (error)
-    {
-        sLog.outError("Player::_SaveInventory - one or more errors occurred save aborted!");
-        ChatHandler(this).SendSysMessage(LANG_ITEM_SAVE_FAILED);
-        return;
-    }
-
-    static SqlStatementID insertInventory ;
-    static SqlStatementID updateInventory ;
-    static SqlStatementID deleteInventory ;
-
-    for (size_t i = 0; i < m_itemUpdateQueue.size(); ++i)
-    {
-        Item* item = m_itemUpdateQueue[i];
-        if (!item)
-        {
-            continue;
-        }
-
-        Bag* container = item->GetContainer();
-        uint32 bag_guid = container ? container->GetGUIDLow() : 0;
-
-        switch (item->GetState())
-        {
-            case ITEM_NEW:
-            {
-                SqlStatement stmt = CharacterDatabase.CreateStatement(insertInventory, "INSERT INTO `character_inventory` (`guid`,`bag`,`slot`,`item`,`item_template`) VALUES (?, ?, ?, ?, ?)");
-                stmt.addUInt32(GetGUIDLow());
-                stmt.addUInt32(bag_guid);
-                stmt.addUInt8(item->GetSlot());
-                stmt.addUInt32(item->GetGUIDLow());
-                stmt.addUInt32(item->GetEntry());
-                stmt.Execute();
-            }
-            break;
-            case ITEM_CHANGED:
-            {
-                SqlStatement stmt = CharacterDatabase.CreateStatement(updateInventory, "UPDATE `character_inventory` SET `guid` = ?, `bag` = ?, `slot` = ?, `item_template` = ? WHERE `item` = ?");
-                stmt.addUInt32(GetGUIDLow());
-                stmt.addUInt32(bag_guid);
-                stmt.addUInt8(item->GetSlot());
-                stmt.addUInt32(item->GetEntry());
-                stmt.addUInt32(item->GetGUIDLow());
-                stmt.Execute();
-            }
-            break;
-            case ITEM_REMOVED:
-            {
-                SqlStatement stmt = CharacterDatabase.CreateStatement(deleteInventory, "DELETE FROM `character_inventory` WHERE `item` = ?");
-                stmt.PExecute(item->GetGUIDLow());
-            }
-            break;
-            case ITEM_UNCHANGED:
-                break;
-        }
-
-        item->SaveToDB();                                   // item have unchanged inventory record and can be save standalone
-    }
-    m_itemUpdateQueue.clear();
-}
-
-/**
- * @brief Saves changed mail records and deletes removed mail from the database.
- */
-void Player::_SaveMail()
-{
-    static SqlStatementID updateMail ;
-    static SqlStatementID deleteMailItems ;
-
-    static SqlStatementID deleteItem ;
-    static SqlStatementID deleteMain ;
-    static SqlStatementID deleteItems ;
-
-    for (PlayerMails::iterator itr = m_mail.begin(); itr != m_mail.end(); ++itr)
-    {
-        Mail* m = (*itr);
-        if (m->state == MAIL_STATE_CHANGED)
-        {
-            SqlStatement stmt = CharacterDatabase.CreateStatement(updateMail, "UPDATE `mail` SET `has_items` = ?, `expire_time` = ?, `deliver_time` = ?, `money` = ?, `cod` = ?, `checked` = ? WHERE `id` = ?");
-            stmt.addUInt32(m->HasItems() ? 1 : 0);
-            stmt.addUInt64(uint64(m->expire_time));
-            stmt.addUInt64(uint64(m->deliver_time));
-            stmt.addUInt32(m->money);
-            stmt.addUInt32(m->COD);
-            stmt.addUInt32(m->checked);
-            stmt.addUInt32(m->messageID);
-            stmt.Execute();
-
-            if (!m->removedItems.empty())
-            {
-                stmt = CharacterDatabase.CreateStatement(deleteMailItems, "DELETE FROM `mail_items` WHERE `item_guid` = ?");
-
-                for (std::vector<uint32>::const_iterator itr2 = m->removedItems.begin(); itr2 != m->removedItems.end(); ++itr2)
-                {
-                    stmt.PExecute(*itr2);
-                }
-
-                m->removedItems.clear();
-            }
-            m->state = MAIL_STATE_UNCHANGED;
-        }
-        else if (m->state == MAIL_STATE_DELETED)
-        {
-            if (m->HasItems())
-            {
-                SqlStatement stmt = CharacterDatabase.CreateStatement(deleteItem, "DELETE FROM `item_instance` WHERE `guid` = ?");
-                for (MailItemInfoVec::const_iterator itr2 = m->items.begin(); itr2 != m->items.end(); ++itr2)
-                {
-                    stmt.PExecute(itr2->item_guid);
-                }
-            }
-
-            SqlStatement stmt = CharacterDatabase.CreateStatement(deleteMain, "DELETE FROM `mail` WHERE `id` = ?");
-            stmt.PExecute(m->messageID);
-
-            stmt = CharacterDatabase.CreateStatement(deleteItems, "DELETE FROM `mail_items` WHERE `mail_id` = ?");
-            stmt.PExecute(m->messageID);
-        }
-    }
-
-    // deallocate deleted mails...
-    for (PlayerMails::iterator itr = m_mail.begin(); itr != m_mail.end();)
-    {
-        if ((*itr)->state == MAIL_STATE_DELETED)
-        {
-            Mail* m = *itr;
-            m_mail.erase(itr);
-            delete m;
-            itr = m_mail.begin();
-        }
-        else
-        {
-            ++itr;
-        }
-    }
-
-    m_mailsUpdated = false;
-}
-
-/**
- * @brief Saves tracked quest status progress to the database.
- */
-void Player::_SaveQuestStatus()
-{
-    static SqlStatementID insertQuestStatus ;
-
-    static SqlStatementID updateQuestStatus ;
-
-    // we don't need transactions here.
-    for (QuestStatusMap::iterator i = mQuestStatus.begin(); i != mQuestStatus.end(); ++i)
-    {
-        QuestStatusData &questStatus = i->second;
-        switch (questStatus.uState)
-        {
-            case QUEST_NEW :
-            {
-                SqlStatement stmt = CharacterDatabase.CreateStatement(insertQuestStatus, "INSERT INTO `character_queststatus` (`guid`,`quest`,`status`,`rewarded`,`explored`,`timer`,`mobcount1`,`mobcount2`,`mobcount3`,`mobcount4`,`itemcount1`,`itemcount2`,`itemcount3`,`itemcount4`,`itemcount5`,`itemcount6`) "
-                                    "VALUES (?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?)");
-
-                stmt.addUInt32(GetGUIDLow());
-                stmt.addUInt32(i->first);
-                stmt.addUInt8(questStatus.m_status);
-                stmt.addUInt8(questStatus.m_rewarded);
-                stmt.addUInt8(questStatus.m_explored);
-                stmt.addUInt64(uint64(questStatus.m_timer / IN_MILLISECONDS + sWorld.GetGameTime()));
-                for (int k = 0; k < QUEST_OBJECTIVES_COUNT; ++k)
-                {
-                    stmt.addUInt32(questStatus.m_creatureOrGOcount[k]);
-                }
-                for (int k = 0; k < QUEST_ITEM_OBJECTIVES_COUNT; ++k)
-                {
-                    stmt.addUInt32(questStatus.m_itemcount[k]);
-                }
-                stmt.Execute();
-            }
-            break;
-            case QUEST_CHANGED :
-            {
-                SqlStatement stmt = CharacterDatabase.CreateStatement(updateQuestStatus, "UPDATE `character_queststatus` SET `status` = ?,`rewarded` = ?,`explored` = ?,`timer` = ?,"
-                                    "`mobcount1` = ?,`mobcount2` = ?,`mobcount3` = ?,`mobcount4` = ?,`itemcount1` = ?,`itemcount2` = ?,`itemcount3` = ?,`itemcount4` = ?,`itemcount5` = ?,`itemcount6` = ? WHERE `guid` = ? AND `quest` = ?");
-
-                stmt.addUInt8(questStatus.m_status);
-                stmt.addUInt8(questStatus.m_rewarded);
-                stmt.addUInt8(questStatus.m_explored);
-                stmt.addUInt64(uint64(questStatus.m_timer / IN_MILLISECONDS + sWorld.GetGameTime()));
-                for (int k = 0; k < QUEST_OBJECTIVES_COUNT; ++k)
-                {
-                    stmt.addUInt32(questStatus.m_creatureOrGOcount[k]);
-                }
-                for (int k = 0; k < QUEST_ITEM_OBJECTIVES_COUNT; ++k)
-                {
-                    stmt.addUInt32(questStatus.m_itemcount[k]);
-                }
-                stmt.addUInt32(GetGUIDLow());
-                stmt.addUInt32(i->first);
-                stmt.Execute();
-            }
-            break;
-            case QUEST_UNCHANGED:
-                break;
-        };
-        questStatus.uState = QUEST_UNCHANGED;
-    }
-}
-
-void Player::_SaveDailyQuestStatus()
-{
-    if (!m_DailyQuestChanged)
-    {
-        return;
-    }
-
-    // we don't need transactions here.
-    static SqlStatementID delQuestStatus ;
-    static SqlStatementID insQuestStatus ;
-
-    SqlStatement stmtDel = CharacterDatabase.CreateStatement(delQuestStatus, "DELETE FROM `character_queststatus_daily` WHERE `guid` = ?");
-    SqlStatement stmtIns = CharacterDatabase.CreateStatement(insQuestStatus, "INSERT INTO `character_queststatus_daily` (`guid`,`quest`) VALUES (?, ?)");
-
-    stmtDel.PExecute(GetGUIDLow());
-
-    for (uint32 quest_daily_idx = 0; quest_daily_idx < PLAYER_MAX_DAILY_QUESTS; ++quest_daily_idx)
-    {
-        if (GetUInt32Value(PLAYER_FIELD_DAILY_QUESTS_1 + quest_daily_idx))
-        {
-            stmtIns.PExecute(GetGUIDLow(), GetUInt32Value(PLAYER_FIELD_DAILY_QUESTS_1 + quest_daily_idx));
-        }
-    }
-    m_DailyQuestChanged = false;
-}
-
-void Player::_SaveWeeklyQuestStatus()
-{
-    if (!m_WeeklyQuestChanged || m_weeklyquests.empty())
-    {
-        return;
-    }
-
-    // we don't need transactions here.
-    static SqlStatementID delQuestStatus ;
-    static SqlStatementID insQuestStatus  ;
-
-    SqlStatement stmtDel = CharacterDatabase.CreateStatement(delQuestStatus, "DELETE FROM `character_queststatus_weekly` WHERE `guid` = ?");
-    SqlStatement stmtIns =  CharacterDatabase.CreateStatement(insQuestStatus, "INSERT INTO `character_queststatus_weekly` (`guid`,`quest`) VALUES (?, ?)");
-
-    stmtDel.PExecute(GetGUIDLow());
-
-    for (QuestSet::const_iterator iter = m_weeklyquests.begin(); iter != m_weeklyquests.end(); ++iter)
-    {
-        uint32 quest_id  = *iter;
-        stmtIns.PExecute(GetGUIDLow(), quest_id);
-    }
-
-    m_WeeklyQuestChanged = false;
-}
-
-void Player::_SaveMonthlyQuestStatus()
-{
-    if (!m_MonthlyQuestChanged || m_monthlyquests.empty())
-    {
-        return;
-    }
-
-    // we don't need transactions here.
-    static SqlStatementID deleteQuest ;
-    static SqlStatementID insertQuest ;
-
-    SqlStatement stmtDel = CharacterDatabase.CreateStatement(deleteQuest, "DELETE FROM `character_queststatus_monthly` WHERE `guid` = ?");
-    SqlStatement stmtIns = CharacterDatabase.CreateStatement(insertQuest, "INSERT INTO `character_queststatus_monthly` (`guid`, `quest`) VALUES (?, ?)");
-
-    stmtDel.PExecute(GetGUIDLow());
-
-    for (QuestSet::const_iterator iter = m_monthlyquests.begin(); iter != m_monthlyquests.end(); ++iter)
-    {
-        uint32 quest_id = *iter;
-        stmtIns.PExecute(GetGUIDLow(), quest_id);
-    }
-
-    m_MonthlyQuestChanged = false;
-}
-
-/**
- * @brief Saves skill value changes to the database.
- */
-void Player::_SaveSkills()
-{
-    static SqlStatementID delSkills ;
-    static SqlStatementID insSkills ;
-    static SqlStatementID updSkills ;
-
-    // we don't need transactions here.
-    for (SkillStatusMap::iterator itr = mSkillStatus.begin(); itr != mSkillStatus.end();)
-    {
-        if (itr->second.uState == SKILL_UNCHANGED)
-        {
-            ++itr;
-            continue;
-        }
-
-        if (itr->second.uState == SKILL_DELETED)
-        {
-            SqlStatement stmt = CharacterDatabase.CreateStatement(delSkills, "DELETE FROM `character_skills` WHERE `guid` = ? AND `skill` = ?");
-            stmt.PExecute(GetGUIDLow(), itr->first);
-            mSkillStatus.erase(itr++);
-            continue;
-        }
-
-        uint16 field = itr->second.pos / 2;
-        uint8 offset = itr->second.pos & 1;
-
-        uint16 value = GetUInt16Value(PLAYER_SKILL_RANK_0 + field, offset);
-        uint16 max = GetUInt16Value(PLAYER_SKILL_MAX_RANK_0 + field, offset);
-
-        switch (itr->second.uState)
-        {
-            case SKILL_NEW:
-            {
-                SqlStatement stmt = CharacterDatabase.CreateStatement(insSkills, "INSERT INTO `character_skills` (`guid`, `skill`, `value`, `max`) VALUES (?, ?, ?, ?)");
-                stmt.PExecute(GetGUIDLow(), itr->first, value, max);
-            }
-            break;
-            case SKILL_CHANGED:
-            {
-                SqlStatement stmt = CharacterDatabase.CreateStatement(updSkills, "UPDATE `character_skills` SET `value` = ?, `max` = ? WHERE `guid` = ? AND `skill` = ?");
-                stmt.PExecute(value, max, GetGUIDLow(), itr->first);
-            }
-            break;
-            case SKILL_UNCHANGED:
-            case SKILL_DELETED:
-                MANGOS_ASSERT(false);
-                break;
-        };
-        itr->second.uState = SKILL_UNCHANGED;
-
-        ++itr;
-    }
-}
-
-/**
- * @brief Saves learned spell state changes to the database.
- */
-void Player::_SaveSpells()
-{
-    static SqlStatementID delSpells ;
-    static SqlStatementID insSpells ;
-
-    SqlStatement stmtDel = CharacterDatabase.CreateStatement(delSpells, "DELETE FROM `character_spell` WHERE `guid` = ? and `spell` = ?");
-    SqlStatement stmtIns = CharacterDatabase.CreateStatement(insSpells, "INSERT INTO `character_spell` (`guid`,`spell`,`active`,`disabled`) VALUES (?, ?, ?, ?)");
-
-    for (PlayerSpellMap::iterator itr = m_spells.begin(), next = m_spells.begin(); itr != m_spells.end();)
-    {
-        uint32 talentCosts = GetTalentSpellCost(itr->first);
-
-        if (!talentCosts)
-        {
-            if (itr->second.state == PLAYERSPELL_REMOVED || itr->second.state == PLAYERSPELL_CHANGED)
-            {
-                stmtDel.PExecute(GetGUIDLow(), itr->first);
-            }
-
-            // add only changed/new not dependent spells
-            if (!itr->second.dependent && (itr->second.state == PLAYERSPELL_NEW || itr->second.state == PLAYERSPELL_CHANGED))
-            {
-                stmtIns.PExecute(GetGUIDLow(), itr->first, uint8(itr->second.active ? 1 : 0), uint8(itr->second.disabled ? 1 : 0));
-            }
-        }
-
-        if (itr->second.state == PLAYERSPELL_REMOVED)
-        {
-            m_spells.erase(itr++);
-        }
-        else
-        {
-            itr->second.state = PLAYERSPELL_UNCHANGED;
-            ++itr;
-        }
-    }
-}
-
-void Player::_SaveTalents()
-{
-    static SqlStatementID delTalents ;
-    static SqlStatementID insTalents ;
-
-    SqlStatement stmtDel = CharacterDatabase.CreateStatement(delTalents, "DELETE FROM `character_talent` WHERE `guid` = ? and `talent_id` = ? and `spec` = ?");
-    SqlStatement stmtIns = CharacterDatabase.CreateStatement(insTalents, "INSERT INTO `character_talent` (`guid`, `talent_id`, `current_rank`, `spec`) VALUES (?, ?, ?, ?)");
-
-    for (uint32 i = 0; i < MAX_TALENT_SPEC_COUNT; ++i)
-    {
-        for (PlayerTalentMap::iterator itr = m_talents[i].begin(); itr != m_talents[i].end();)
-        {
-            if (itr->second.state == PLAYERSPELL_REMOVED || itr->second.state == PLAYERSPELL_CHANGED)
-            {
-                stmtDel.PExecute(GetGUIDLow(), itr->first, i);
-            }
-
-            // add only changed/new talents
-            if (itr->second.state == PLAYERSPELL_NEW || itr->second.state == PLAYERSPELL_CHANGED)
-            {
-                stmtIns.PExecute(GetGUIDLow(), itr->first, itr->second.currentRank, i);
-            }
-
-            if (itr->second.state == PLAYERSPELL_REMOVED)
-            {
-                m_talents[i].erase(itr++);
-            }
-            else
-            {
-                itr->second.state = PLAYERSPELL_UNCHANGED;
-                ++itr;
-            }
-        }
-    }
-}
-
-// save player stats -- only for external usage
-// real stats will be recalculated on player login
-void Player::_SaveStats()
-{
-    // check if stat saving is enabled and if char level is high enough
-    if (!sWorld.getConfig(CONFIG_UINT32_MIN_LEVEL_STAT_SAVE) || getLevel() < sWorld.getConfig(CONFIG_UINT32_MIN_LEVEL_STAT_SAVE))
-    {
-        return;
-    }
-
-    static SqlStatementID delStats ;
-    static SqlStatementID insertStats ;
-
-    SqlStatement stmt = CharacterDatabase.CreateStatement(delStats, "DELETE FROM `character_stats` WHERE `guid` = ?");
-    stmt.PExecute(GetGUIDLow());
-
-    stmt = CharacterDatabase.CreateStatement(insertStats, "INSERT INTO `character_stats` (`guid`, `maxhealth`, `maxpower1`, `maxpower2`, `maxpower3`, `maxpower4`, `maxpower5`,"
-            "`strength`, `agility`, `stamina`, `intellect`, `spirit`, `armor`, `resHoly`, `resFire`, `resNature`, `resFrost`, `resShadow`, `resArcane`, "
-            "`blockPct`, `dodgePct`, `parryPct`, `critPct`, `rangedCritPct`, `spellCritPct`, `attackPower`, `rangedAttackPower`, `spellPower`) "
-        "VALUES ( ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?)");
-
-    stmt.addUInt32(GetGUIDLow());
-    stmt.addUInt32(GetMaxHealth());
-    for (uint32 i = 0; i < MAX_STORED_POWERS; ++i)
-    {
-        stmt.addUInt32(GetMaxPowerByIndex(i));
-    }
-    for (int i = 0; i < MAX_STATS; ++i)
-    {
-        stmt.addFloat(GetStat(Stats(i)));
-    }
-    // armor + school resistances
-    for (int i = 0; i < MAX_SPELL_SCHOOL; ++i)
-    {
-        stmt.addUInt32(GetResistance(SpellSchools(i)));
-    }
-    stmt.addFloat(GetFloatValue(PLAYER_BLOCK_PERCENTAGE));
-    stmt.addFloat(GetFloatValue(PLAYER_DODGE_PERCENTAGE));
-    stmt.addFloat(GetFloatValue(PLAYER_PARRY_PERCENTAGE));
-    stmt.addFloat(GetFloatValue(PLAYER_CRIT_PERCENTAGE));
-    stmt.addFloat(GetFloatValue(PLAYER_RANGED_CRIT_PERCENTAGE));
-    stmt.addFloat(GetFloatValue(PLAYER_SPELL_CRIT_PERCENTAGE1));
-    stmt.addUInt32(GetUInt32Value(UNIT_FIELD_ATTACK_POWER));
-    stmt.addUInt32(GetUInt32Value(UNIT_FIELD_RANGED_ATTACK_POWER));
-    stmt.addUInt32(GetBaseSpellPowerBonus());
-
-    stmt.Execute();
-}
 
 /**
  * @brief Writes the player's current combat and stat values to the debug log.
@@ -12497,28 +11573,6 @@ void Player::SendAttackSwingNotInRange()
 {
     WorldPacket data(SMSG_ATTACKSWING_NOTINRANGE, 0);
     GetSession()->SendPacket(&data);
-}
-
-/**
- * @brief Writes a character position directly to the database.
- *
- * @param guid The character GUID to update.
- * @param mapid The destination map identifier.
- * @param x The X coordinate.
- * @param y The Y coordinate.
- * @param z The Z coordinate.
- * @param o The orientation.
- * @param zone The zone identifier.
- */
-void Player::SavePositionInDB(ObjectGuid guid, uint32 mapid, float x, float y, float z, float o, uint32 zone)
-{
-    std::ostringstream ss;
-    ss << "UPDATE `characters` SET `position_x`='" << x << "',`position_y`='" << y
-       << "',`position_z`='" << z << "',`orientation`='" << o << "',`map`='" << mapid
-       << "',`zone`='" << zone << "',`trans_x`='0',`trans_y`='0',`trans_z`='0',"
-       << "`transguid`='0',`taxi_path`='' WHERE `guid`='" << guid.GetCounter() << "'";
-    DEBUG_LOG("%s", ss.str().c_str());
-    CharacterDatabase.Execute(ss.str().c_str());
 }
 
 /**
@@ -18660,118 +17714,6 @@ void Player::SetEquipmentSet(uint32 index, EquipmentSet eqset)
     }
 
     eqslot.state = old_state == EQUIPMENT_SET_NEW ? EQUIPMENT_SET_NEW : EQUIPMENT_SET_CHANGED;
-}
-
-void Player::_SaveEquipmentSets()
-{
-    static SqlStatementID updSets ;
-    static SqlStatementID insSets ;
-    static SqlStatementID delSets ;
-
-    for (EquipmentSets::iterator itr = m_EquipmentSets.begin(); itr != m_EquipmentSets.end();)
-    {
-        uint32 index = itr->first;
-        EquipmentSet& eqset = itr->second;
-        switch (eqset.state)
-        {
-            case EQUIPMENT_SET_UNCHANGED:
-                ++itr;
-                break;                                      // nothing do
-            case EQUIPMENT_SET_CHANGED:
-            {
-                SqlStatement stmt = CharacterDatabase.CreateStatement(updSets, "UPDATE `character_equipmentsets` SET `name`=?, `iconname`=?, `ignore_mask`=?, `item0`=?, `item1`=?, `item2`=?, `item3`=?, `item4`=?, "
-                                    "`item5`=?, `item6`=?, `item7`=?, `item8`=?, `item9`=?, `item10`=?, `item11`=?, `item12`=?, `item13`=?, `item14`=?, "
-                                    "`item15`=?, `item16`=?, `item17`=?, `item18`=? WHERE `guid`=? AND `setguid`=? AND `setindex`=?");
-
-                stmt.addString(eqset.Name);
-                stmt.addString(eqset.IconName);
-                stmt.addUInt32(eqset.IgnoreMask);
-
-                for (int i = 0; i < EQUIPMENT_SLOT_END; ++i)
-                {
-                    stmt.addUInt32(eqset.Items[i]);
-                }
-
-                stmt.addUInt32(GetGUIDLow());
-                stmt.addUInt64(eqset.Guid);
-                stmt.addUInt32(index);
-
-                stmt.Execute();
-
-                eqset.state = EQUIPMENT_SET_UNCHANGED;
-                ++itr;
-                break;
-            }
-            case EQUIPMENT_SET_NEW:
-            {
-                SqlStatement stmt = CharacterDatabase.CreateStatement(insSets, "INSERT INTO `character_equipmentsets` VALUES (?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?)");
-                stmt.addUInt32(GetGUIDLow());
-                stmt.addUInt64(eqset.Guid);
-                stmt.addUInt32(index);
-                stmt.addString(eqset.Name);
-                stmt.addString(eqset.IconName);
-                stmt.addUInt32(eqset.IgnoreMask);
-
-                for (int i = 0; i < EQUIPMENT_SLOT_END; ++i)
-                {
-                    stmt.addUInt32(eqset.Items[i]);
-                }
-
-                stmt.Execute();
-
-                eqset.state = EQUIPMENT_SET_UNCHANGED;
-                ++itr;
-                break;
-            }
-            case EQUIPMENT_SET_DELETED:
-            {
-                SqlStatement stmt = CharacterDatabase.CreateStatement(delSets, "DELETE FROM `character_equipmentsets` WHERE `setguid` = ?");
-                stmt.PExecute(eqset.Guid);
-                m_EquipmentSets.erase(itr++);
-                break;
-            }
-        }
-    }
-}
-
-/**
- * @brief Saves battleground return position and instance data to the database.
- */
-void Player::_SaveBGData()
-{
-    // nothing save
-    if (!m_bgData.m_needSave)
-    {
-        return;
-    }
-
-    static SqlStatementID delBGData ;
-    static SqlStatementID insBGData ;
-
-    SqlStatement stmt =  CharacterDatabase.CreateStatement(delBGData, "DELETE FROM `character_battleground_data` WHERE `guid` = ?");
-
-    stmt.PExecute(GetGUIDLow());
-
-    if (m_bgData.bgInstanceID)
-    {
-        stmt = CharacterDatabase.CreateStatement(insBGData, "INSERT INTO `character_battleground_data` VALUES (?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?)");
-        /* guid, bgInstanceID, bgTeam, x, y, z, o, map, taxi[0], taxi[1], mountSpell */
-        stmt.addUInt32(GetGUIDLow());
-        stmt.addUInt32(m_bgData.bgInstanceID);
-        stmt.addUInt32(uint32(m_bgData.bgTeam));
-        stmt.addFloat(m_bgData.joinPos.coord_x);
-        stmt.addFloat(m_bgData.joinPos.coord_y);
-        stmt.addFloat(m_bgData.joinPos.coord_z);
-        stmt.addFloat(m_bgData.joinPos.orientation);
-        stmt.addUInt32(m_bgData.joinPos.mapid);
-        stmt.addUInt32(m_bgData.taxiPath[0]);
-        stmt.addUInt32(m_bgData.taxiPath[1]);
-        stmt.addUInt32(m_bgData.mountSpell);
-
-        stmt.Execute();
-    }
-
-    m_bgData.m_needSave = false;
 }
 
 void Player::DeleteEquipmentSet(uint64 setGuid)
